requirements.txt
openpyxl==3.0.10
plotly==4.14.3
pandas==1.4.3
streamlit==1.12.0
kafka-python